+++ RetailStore/Extensions/OrderExtensions.cs
@@ -4,7 +4,7 @@ namespace RetailStore.Extentions;
 
 public static class OrderExtensions
 {
-    public static double TotalRevenue(this double price, int quantity) 
+    public static decimal TotalRevenue(this decimal price, int quantity) 
     {
         return price * quantity;
     }
